+++ lib/pechkin/message_matcher.rb
@@ -10,6 +10,13 @@ module Pechkin
   class MessageMatcher
     KEY_ALLOW = 'allow'.freeze
     KEY_FORBID = 'forbid'.freeze
+
+    attr_reader :logger
+
+    def initialize(logger)
+      @logger = logger
+    end
+
     # Checks data object against allow / forbid rule sets in message
     # configuration. If data object matches rules it means we can process this
     # data and send message.
@@ -23,10 +30,10 @@ module Pechkin
 
       if message_config.key?(KEY_ALLOW)
         rules = message_config[KEY_ALLOW]
-        rules.any? { |r| check_rule(r, data) }
+        rules.any? { |r| check_rule(r, r, data) }
       elsif message_config.key?(KEY_FORBID)
         rules = message_config[KEY_FORBID]
-        rules.all? { |r| !check_rule(r, data) }
+        rules.all? { |r| !check_rule(r, r, data) }
       else
         true
       end
@@ -43,35 +50,42 @@ module Pechkin
     # Check rule object against data. Rules are checked recursievly, i.e. we
     # can take one field in rule and check it separately as new rule. If all
     # fields are passed check then whole rule passed.
-    def check_rule(rule, data)
-      if rule.is_a?(Hash)
-        check_hash_rule(rule, data)
-      elsif rule.is_a?(Array)
-        check_array_rule(rule, data)
-      elsif rule.is_a?(String)
-        check_string_rule(rule, data)
-      else
-        rule.eql?(data)
+    def check_rule(top_rule, sub_rule, data)
+      result = if sub_rule.is_a?(Hash)
+                 check_hash_rule(top_rule, sub_rule, data)
+               elsif sub_rule.is_a?(Array)
+                 check_array_rule(top_rule, sub_rule, data)
+               elsif sub_rule.is_a?(String)
+                 check_string_rule(top_rule, sub_rule, data)
+               else
+                 sub_rule.eql?(data)
+               end
+
+      unless result
+        logger.info "Expected #{sub_rule.to_json} got #{data.to_json} when" \
+                    " checking #{top_rule.to_json}"
       end
+
+      result
     end
 
-    def check_hash_rule(hash, data)
+    def check_hash_rule(top_rule, hash, data)
       return false unless data.is_a?(Hash)
 
       hash.all? do |key, value|
-        data.key?(key) && check_rule(value, data[key])
+        data.key?(key) && check_rule(top_rule, value, data[key])
       end
     end
 
-    def check_array_rule(array, data)
+    def check_array_rule(top_rule, array, data)
       return false unless data.is_a?(Array)
 
       # Deep array check needs to be done against all elements so we zip arrays
       # to pair each rule with data element
-      array.zip(data).all? { |r, d| check_rule(r, d) }
+      array.zip(data).all? { |r, d| check_rule(top_rule, r, d) }
     end
 
-    def check_string_rule(str, data)
+    def check_string_rule(_top_rule, str, data)
       str.eql? data
     end
   end